+++ e2e/cypress/e2e/interactive_GPX.cy.js
@@ -1,0 +1,15 @@
+import {export_with_interaction, test_and_save_download} from './utils.js';
+
+before(() => {
+    cy.exec('rm -rf cypress/downloads/*');
+});
+
+
+it('Test small GPX file with UI change', () => {
+
+    const file = "cypress/fixtures/test_small.gpx";
+    export_with_interaction(file);
+    test_and_save_download(file);
+
+})
+