+++ docs/changelog/changelog.md
@@ -2,7 +2,22 @@
 
 A complete, technical changelog is available at [GitHub](https://github.com/cevi/automatic_walk-time_tables/releases).
 
-## Version 3.0.0
+## Version 3 - Interactive Exports
+- Interactive exports: modify the selected waypoints using points of interest (POI) (3.0.0)
+- Improved Documentation and Changelog (on this page) (3.0.0)
+- combine all maps in a single PDF file (3.0.0)
+- Add download button (for cases where the auto download does not work) (3.1.0)
+- Bug fixes
+- Dependency updates
 
-- Interactive Exports
-- fix ...
+## Version 2 - Web App
+- Web app with user guide
+- Improved point selecting algorithm (2.3.0)
+- Change in background map (2.3.0)
+- Support for KML files (2.2.0)
+- Bug fixes
+- Dependency updates
+
+## Version 1 - Initial Script
+- Docker-based deployment
+- Support for GPX files